perch-agent/pyproject.toml
[project]
name = "perch-agent"
version = "0.1.0"
description = "OpenChoreo control-plane copilot — chat over the openchoreo MCP with confirmable writes"
requires-python = ">=3.14"
dependencies = [
    "authlib>=1.6.8",
    "fastapi>=0.133.1",
    "httpx>=0.23.0",
    "jinja2>=3.1.0",
    "langchain>=1.0.5",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-openai>=1.1.10",
    "pyjwt[crypto]>=2.8.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[tool.uv]
exclude-dependencies = ["zstandard", "pywin32"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.ruff]
line-length = 100
target-version = "py314"
extend-exclude = [".venv", "__pycache__", ".git"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
    "B904",  # raise from
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
